Python/requirements.txt
opencv-python
PySimpleGUI
numpy
pyserial
